:recycle: Extract Setup function and Template struct to setup.go
