+++ src/admin/auth_handler.go
@@ -2,8 +2,6 @@ package admin
 
 import (
 	"fmt"
-	"html/template"
-	"io"
 	"net/http"
 	"os"
 	"time"
@@ -18,37 +16,6 @@ import (
 )
 
 type AuthHandler struct{}
-
-var e *echo.Echo
-
-func Setup(echo *echo.Echo, dir string) {
-	e = echo
-
-	h := &AuthHandler{}
-	t := &Template{
-		templates: template.Must(template.ParseGlob(dir + "/*.html")),
-	}
-	e.Renderer = t
-
-	g := e.Group("/admin/auths")
-	g.GET("", h.withFlash(h.index))
-	g.POST("", h.withFlash(h.create))
-	g.POST("/:id/disable", h.AuthHandler(h.disable))
-	g.POST("/:id/delete", h.AuthHandler(h.destroy))
-}
-
-type Template struct {
-	templates *template.Template
-}
-
-func (t *Template) Render(w io.Writer, name string, data interface{}, c echo.Context) error {
-	return t.templates.ExecuteTemplate(w, name, data)
-}
-
-type Flash struct {
-	Alert  string
-	Notice string
-}
 
 func (h *AuthHandler) setFlash(c echo.Context, name, value string) {
 	h.setFlashWithExpire(c, name, value, time.Now().Add(10*time.Minute))
